+++ CHANGELOG.md
@@ -158,6 +158,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   text when outputting to a non-terminal, instead of emitting raw OSC 8 escape
   codes. [#7592](https://github.com/jj-vcs/jj/issues/7592)
 
+* Conditional configuration based on environment variables with `--when.environments`.
+  [#8779](https://github.com/jj-vcs/jj/pull/8779)
+
 ### Fixed bugs
 
 * `jj git init --colocate` now refuses to run inside a Git worktree, providing